% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_conductance.R
\name{aerodynamic_conductance_BM2013}
\alias{aerodynamic_conductance_BM2013}
\title{Calculate different bulk conductance estimates}
\usage{
aerodynamic_conductance_BM2013(ustar, u, karman = 0.4, Sc = 1,
  Pr = 1)
}
\description{
Calculate different bulk conductance estimates
}
