% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_conductance.R
\name{aerodynamic_conductance_Hinvert}
\alias{aerodynamic_conductance_Hinvert}
\title{Estimate of the aerodynamic conductance for heat by inverting the bulk conductance formula for sensible heat}
\usage{
aerodynamic_conductance_Hinvert(H, Ts, Ta, rho = 1.2, cp = 1004)
}
\arguments{
\item{H}{sensible heat flux in W m-2}

\item{Ts}{surface temperature in K, ideally this is the surface temperature of the heat source}

\item{Ta}{air temperature in K}

\item{rho}{density of air in kg m-3}

\item{cp}{specific heat of air at constant pressure in J kg-1 K-1}
}
\value{
aerodynamic conductance in m/s
}
\description{
Estimate of the aerodynamic conductance for heat by inverting the bulk conductance formula for sensible heat
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
