% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_conductance.R
\name{aerodynamic_conductance_ustaru}
\alias{aerodynamic_conductance_ustaru}
\title{Estimate of the aerodynamic conductance for momentum from wind speed and friction velocity}
\usage{
aerodynamic_conductance_ustaru(ustar, u)
}
\arguments{
\item{ustar}{friction velocity in m/s}

\item{u}{horizontal wind speed in m/s}
}
\value{
aerodynamic conductance in m/s
}
\description{
Estimate of the aerodynamic conductance for momentum from wind speed and friction velocity
}
\author{
Maik Renner, mrenner@bgc-jena.mpg.de
}
