% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aggregate.R
\name{anomean}
\alias{anomean}
\title{Calc an anomaly given the full data
 
Simple helper fun}
\usage{
anomean(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{further arguments passed to mean(x, ...) such as na.rm=TRUE}
}
\value{
numeric vector of anomalies
}
\description{
Calc an anomaly given the full data
 
Simple helper fun
}
