% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_conductance.R
\name{LatentHeatFlux_PenmanMonteith}
\alias{LatentHeatFlux_PenmanMonteith}
\title{Calculate the Latent Heat flux using the Penman-Monteith formulation}
\usage{
LatentHeatFlux_PenmanMonteith(AE, s, esurf, eair, ga, gs, rho = 1.2,
  cp = 1004, psychro = 0.65)
}
\arguments{
\item{AE}{Available Energy usually Net Radiation - Soil Heat flux in W m-2}

\item{s}{slope of the saturation vapor pressure curve in hPa/K}

\item{esurf}{surface water vapour pressure hPa}

\item{eair}{vapour pressure in air hPa}

\item{ga}{aerodynamic conductance for water vapor in m/s}

\item{gs}{surface / or stomatal conductance for water vapor in m/s}

\item{rho}{Air density kg/m3}

\item{cp}{heat capacity of air J/(kg K)}

\item{psychro}{psychrometric constant hPa/K}
}
\value{
Latent Heat flux in W/m2
}
\description{
Calculate the Latent Heat flux using the Penman-Monteith formulation
}
\seealso{
[aerodynamic_conductance_withcanopy_Thom1975()] ga needs a leaf boundary layer conductance (also known as excess resistance)
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
