% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vapor.R
\name{MAGNUS}
\alias{MAGNUS}
\title{Empirical equation for saturation water vapor pressure}
\usage{
MAGNUS(air_temp)
}
\arguments{
\item{air_temp}{Temperature in °C}
}
\value{
water vapor pressure at saturation over water surfaces in hPa
}
\description{
Empirical equation for saturation water vapor pressure
}
\details{
separates vapor pressure over water and ice

wikipedia reports different coefficients
}
