% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vapor.R
\name{Magnus_Alduchov1996Improved}
\alias{Magnus_Alduchov1996Improved}
\title{Calculate meteorological variables related to water vapor}
\usage{
Magnus_Alduchov1996Improved(air_temp, c1 = 6.1094, c2 = 17.625,
  c3 = 243.04, ice = FALSE)
}
\arguments{
\item{air_temp}{Temperature in °C.}

\item{c1, c2, c3}{Empirical coefficients with defaults set to values by
Alduchov and Eskridge (1996) Journal of Apllied Meteorology.}

\item{ice}{Logical parameter which switches to the ice params
when an ice surface is there and T < 0 (default ice = FALSE).}
}
\value{
water vapor pressure at saturation over water surfaces or ice surface in hPa
}
\description{
Calculate meteorological variables related to water vapor
}
\details{
Alduchov and Eskridge (1996) Journal of Apllied Meteorology eqs 21 for water and 23 for ice

also provide enhancement factors for moist air (not included here)

see discussion by Koutsoyannis 2012 on the derivation
}
\seealso{
[Magnus_Alduchov1996Improved_slope_sat]
}
\author{
Maik Renner mrenner@bgc-jena.mpg.de
}
