% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vapor.R
\name{Magnus_Alduchov1996Improved_slope_sat}
\alias{Magnus_Alduchov1996Improved_slope_sat}
\title{calc the derivative of the saturation vapor pressure curve of MAGNUS}
\usage{
Magnus_Alduchov1996Improved_slope_sat(air_temp, c2 = 17.625,
  c3 = 243.04, ice = FALSE)
}
\arguments{
\item{air_temp}{Temperature in °C.}

\item{c2, c3}{Empirical coefficients with defaults set to values by
Alduchov and Eskridge (1996) Journal of Apllied Meteorology.}

\item{ice}{Logical parameter which switches to the ice params
when an ice surface is there and T < 0 (default ice = FALSE).}
}
\value{
slope of the saturation water vapor pressure curve hPa K-1.
}
\description{
calc the derivative of the saturation vapor pressure curve of MAGNUS
}
\seealso{
[Magnus_Alduchov1996Improved]
}
