% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.colors.R
\name{makeTransparent}
\alias{makeTransparent}
\title{small utilities functions to deal with colors in R}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{any color code what R understands as character}

\item{alpha}{numeric 0 ... 100 setting the transparancy, 0 no color,
100 full color}
}
\value{
return a color code which can be used in plot(  , col = )
}
\description{
small utilities functions to deal with colors in R
}
\examples{
makeTransparent("green")
makeTransparent("green", alpha = 0.1)
makeTransparent(makeTransparent("green", alpha = 0.1))
plot(0,0, pch = 16, col = "green", cex = 5)
points(0.1,0, pch = 15, col = makeTransparent("green", alpha = 50), cex = 5 )
set.seed(688)
xp = rnorm(7)
yp = rnorm(7)
polygon(xp, yp, border = 3, col = makeTransparent("green", alpha = 80) )
}
\seealso{
<https://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color>
}
\author{
Maik Renner mrenner@bgc-jena.mpg.de
}
