% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aggregate.R
\name{meana}
\alias{meana}
\title{collection of specific useful summary functions}
\usage{
meana(x, na.percentage = 0.5, FUN = mean)
}
\description{
collection of specific useful summary functions
}
\author{
Maik Renner mrenner@bgc-jena.mpg.de
}
