% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aggregate.R
\name{meann}
\alias{meann}
\title{an average mean which requires a minimum of data points nmin
 needed if missing data is missing :)}
\usage{
meann(x, nmin, ...)
}
\arguments{
\item{x}{vector of numeric values}

\item{nmin}{minimum of data points accepted to calculate the mean (integer)}

\item{...}{further arguments to mean() remember especially to set na.rm=TRUE}
}
\description{
an average mean which requires a minimum of data points nmin
 needed if missing data is missing :)
}
\examples{
x = 1:10
meann(x,nmin = 11)
meann(NULL,nmin = 11)
x = c(1,2,NA,Inf,NaN)
meann(x,2,na.rm=TRUE)
x = c(1,2,NA,NA,NaN)
meann(x,2,na.rm=TRUE)
meann(x,3,na.rm=TRUE)
sum(!is.na(x))
}
\author{
Maik Renner mrenner@bgc-jena.mpg.de
}
