% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.table.regression.utils.R
\name{mlm.output.statlong.call}
\alias{mlm.output.statlong.call}
\title{Call a linear regression model and reorganize its output}
\usage{
mlm.output.statlong.call(mula, data, ...)
}
\arguments{
\item{mula}{a formula for a lm() regression provided as character using the column names of the data}

\item{data}{the data table with the column names}

\item{...}{additional arguments to call lm() such as weights and other lm arguments}
}
\value{
a data.table with columns statistic and value
}
\description{
sucessfull error handling for errors on function calls when groups are empty
}
\examples{
DT = as.data.table(mtcars)
DT[, mlm.output.statlong.call(mula = "mpg ~ drat + gear", data = .SD) , by = list(am) ]
# impose an error with a nonexisting column name
DT[, mlm.output.statlong.call(mula = "mpg ~ drat + gear + bogus", data = .SD) , by = list(am) ]
# use ... to weight the regression
DT[, mlm.output.statlong.call(mula = "mpg ~ drat + gear", data = .SD, weight = cyl) , by = list(am) ]
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
