% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaselag.R
\name{phaselag_time}
\alias{phaselag_time}
\title{Calculate the phase lag in time units}
\usage{
phaselag_time(slope1, slope2, nday, timeunitperday)
}
\arguments{
\item{slope1}{Linear regression slope between two variables X and Y.}

\item{slope2}{2nd order regression slope between the variable Y and the time derivative of X dX/dt.}

\item{nday}{Number of measurements per day.}

\item{timeunitperday}{Number of timesteps per day.}
}
\value{
phase lag in time units between two variables.
}
\description{
Helper function which is called by camuffo_phaselag_time()
}
\references{
please cite the following paper when using this code
}
\seealso{
camuffo_phaselag_time
}
