% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_vapor.R
\name{slope_sat}
\alias{slope_sat}
\title{calc the derivative of the saturation vapor pressure curve of MAGNUS}
\usage{
slope_sat(air_temp)
}
\description{
calc the derivative of the saturation vapor pressure curve of MAGNUS
}
