% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aggregate.R
\name{summaryfun_nmin}
\alias{summaryfun_nmin}
\title{a summary function applied to a vector which requires a minimum of data points nmin}
\usage{
summaryfun_nmin(x, nmin, FUN = mean, ...)
}
\description{
needed if missing data is missing :)
add non-missing data lower nmin @20141208
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de
}
