% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aggregate.R
\name{weighted.var}
\alias{weighted.var}
\title{computes an unbiased weigthed variance using reliability weights}
\usage{
weighted.var(x, wt)
}
\description{
computes an unbiased weigthed variance using reliability weights
}
