% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_conductance.R
\name{aerodynamic_conductance_withcanopy_Thom1975}
\alias{aerodynamic_conductance_withcanopy_Thom1975}
\title{combined aerodynamic conductance after Thom 1975 taken from De Kauwe 2017 BG , eq. 3
first term is the turbulent aerodynamic resistance the second term is the canopy boundary layer component}
\usage{
aerodynamic_conductance_withcanopy_Thom1975(ustar, u)
}
\description{
combined aerodynamic conductance after Thom 1975 taken from De Kauwe 2017 BG , eq. 3
first term is the turbulent aerodynamic resistance the second term is the canopy boundary layer component
}
