% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaselag.R
\name{phaselag_time}
\alias{phaselag_time}
\title{Calculate the phase lag in time units}
\usage{
phaselag_time(slope1, slope2, nday, timeunitperday)
}
\arguments{
\item{slope1}{Linear regression slope between two variables X and Y.}

\item{slope2}{2nd order regression slope between the variable Y and the time derivative of X dX/dt.}

\item{nday}{Number of measurements per day.}

\item{timeunitperday}{Number of timesteps per day.}
}
\value{
phase lag in time units between two variables.
}
\description{
Helper function which is called by camuffo_phaselag_time()
}
\references{
please cite the following paper when using this code
Renner, M., Brenner, C., Mallick, K., Wizemann, H.-D., Conte, L., Trebs, I., Wei, J., Wulfmeyer, V., Schulz, K., and Kleidon, A.: Using phase lags to evaluate model biases in simulating the diurnal cycle of evapotranspiration: a case study in Luxembourg, Hydrol. Earth Syst. Sci., 23, 515-535, https://doi.org/10.5194/hess-23-515-2019, 2019.
}
\seealso{
camuffo_phaselag_time
}
\author{
Maik Renner, mrenner [at] bgc-jena.mpg.de, Luigi Conte contributed the harmonic analysis
}
